+++ pkg/components/cnudie/componentresolvers/local.go
@@ -24,8 +24,8 @@ import (
 
 	"github.com/gardener/landscaper/controller-utils/pkg/logging"
 	lc "github.com/gardener/landscaper/controller-utils/pkg/logging/constants"
+	"github.com/gardener/landscaper/pkg/components/model/tar"
 	"github.com/gardener/landscaper/pkg/components/model/types"
-	"github.com/gardener/landscaper/pkg/utils/tar"
 )
 
 // LocalRepositoryType defines the local repository context type.
