+++ pkg/components/cnudie/componentresolvers/oci_test.go
@@ -22,8 +22,8 @@ import (
 	ocispecv1 "github.com/opencontainers/image-spec/specs-go/v1"
 
 	"github.com/gardener/landscaper/controller-utils/pkg/logging"
+	"github.com/gardener/landscaper/pkg/components/model/tar"
 	"github.com/gardener/landscaper/pkg/components/model/types"
-	"github.com/gardener/landscaper/pkg/utils/tar"
 )
 
 func TestConfig(t *testing.T) {
